Add test
git-svn-id: svn://svn.code.sf.net/p/moka-modeller/code@63 578f8393-57f1-403d-81f3-a3b3dc990742

+++ src/lib-gmapkernel/g-map-vertex/gmv-save.cc
@@ -33,7 +33,7 @@ void CGMapVertex::giveNumberToDarts(int ADirectInfoIndex)
 {
   CDynamicCoverageAll it(this);
 
-  for (uint32_t num=0; it.cont(); ++it, ++num)
+  for (unsigned long int num=0; it.cont(); ++it, ++num)
     (*it)->setDirectInfo(ADirectInfoIndex, (void *) num);
 }
 //******************************************************************************
